requirements.txt
numpy
scipy
matplotlib
joblib
cupy-cuda11x; sys_platform == 'linux'  # Optional, depends on CUDA version